pass abbrev_commit to gbr_sha() and move it in gbr_dump_context
Our default sha printing length has been a global variable that's
set when parsing options. As we're moving the default "remote delta"
operation to a separate module, that gets unpleasant to deal with.

So just stash ->abbrev in struct gbr_dump_context and make gbr_sha()
take it as a parameter.